+++ applications/fishing-map/src/features/map/Map.tsx
@@ -26,7 +26,7 @@ import MapInfo from 'features/map/controls/MapInfo'
 import MapControls from 'features/map/controls/MapControls'
 import { selectDebugOptions } from 'features/debug/debug.slice'
 import { formatI18nNumber } from 'features/i18n/i18nNumber'
-import { ClickPopup, HoverPopup } from './Popup'
+import { ClickPopup, HoverPopup } from './popups/Popup'
 import useViewport, { useMapBounds } from './map-viewport.hooks'
 import { useMapboxRef, useMapboxRefReady } from './map.context'
 import styles from './Map.module.css'
